Added support for the Whiley bootpath.
The "bootpath" identifies the location of the Whiley standard library.
At the moment, this is not pulled from Maven central (though it
potentially could be).  Instead, it is just managed as a local
dependency.

The "wy" script defines a JVM property variable "wdk.bootpath" which
identifies the bootpath, based on the install location of the wdk.